Move package and fixes for some issue and missing implementations
