+++ coding/wml-v20/src/test/java/org/n52/sos/decode/WmlObservationProcessDecoderv20Test.java
@@ -38,7 +38,7 @@ import org.n52.sos.ogc.om.values.ReferenceValue;
 import org.n52.sos.ogc.om.values.TextValue;
 import org.n52.sos.ogc.sensorML.SensorMLConstants;
 import org.n52.sos.ogc.sos.SosOffering;
-import org.n52.sos.ogc.wml.ObservationProcess;
+import org.n52.sos.ogc.series.wml.ObservationProcess;
 
 /**
  * Test class for {@link WmlObservationProcessDecoderv20}
