+++ core/api/src/main/java/org/n52/sos/ogc/om/AbstractObservationValue.java
@@ -29,8 +29,8 @@
 package org.n52.sos.ogc.om;
 
 import org.n52.sos.ogc.om.values.Value;
-import org.n52.sos.ogc.wml.DefaultPointMetadata;
-import org.n52.sos.ogc.wml.Metadata;
+import org.n52.sos.ogc.series.wml.DefaultPointMetadata;
+import org.n52.sos.ogc.series.wml.Metadata;
 import org.n52.sos.util.StringHelper;
 
 public abstract class AbstractObservationValue<T extends Value<?>> implements ObservationValue<T> {
