+++ core/api/src/main/java/org/n52/sos/ogc/series/wml/MeasurementTimeseriesMetadata.java
@@ -21,7 +21,7 @@
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA or
  * visit the Free Software Foundation web page, http://www.fsf.org.
  */
-package org.n52.sos.ogc.wml;
+package org.n52.sos.ogc.series.wml;
 
 /**
  * This class implements the OGC WaterML 2.0 element <code>metadata</code>. See <code>/req/uml-timeseries-observation/metadata</code>.
