+++ core/api/src/main/java/org/n52/sos/ogc/series/wml/WaterMLConstants.java
@@ -35,8 +35,11 @@ import org.n52.sos.util.http.MediaType;
 import org.n52.sos.w3c.SchemaLocation;
 
 /**
- * @since 4.0.0
+ * This interface holds all constants required by the OGC WaterML 2.0
+ * model objects.
  * 
+ * @since 4.0.0
+ * @see http://www.opengeospatial.org/standards/waterml
  */
 public interface WaterMLConstants {
     String NS_WML_20 = "http://www.opengis.net/waterml/2.0";
@@ -120,6 +123,8 @@ public interface WaterMLConstants {
     
     String EN_INTERPOLATION_TYPE = "interpolationType";
     
+    String EN_CUMULATIVE = "cumulative";
+
     String EN_UOM = "uom";
     
     QName QN_POINT = new QName(NS_WML_20, EN_POINT, NS_WML_20_PREFIX);
@@ -150,6 +155,106 @@ public interface WaterMLConstants {
 
     QName QN_INTERPOLATION_TYPE = new QName(NS_WML_20, EN_INTERPOLATION_TYPE, NS_WML_20_PREFIX);
 
+    QName QN_CUMULATIVE = new QName(NS_WML_20, EN_CUMULATIVE, NS_WML_20_PREFIX);
+
     QName UOM =  new QName(NS_WML_20, EN_UOM, NS_WML_20_PREFIX);
 
+    /**
+     * @see MeasurementTimeseriesMetadata#isCumulative()
+     */
+    String SERIES_METADATA_CUMULATIVE = NS_WML_20 + "/cumulative";
+
+    String INTERPOLATION_TYPE = "http://www.opengis.net/def/waterml/2.0/interpolationType";
+
+    /**
+     * Hold allowed values for element <code>interpolationType</code>.
+     * 
+     * See <code>/req/xsd-measurement-timeseries-tvp/interpolation-type</code>.
+     * 
+     * @author <a href="mailto:e.h.juerrens@52north.org">Eike Hinderk J&uuml;rrens</a>
+     * @since 4.4.0
+     */
+    public enum InterpolationType {
+        
+        /**
+         * Continuous/Instantaneous
+         * http://www.opengis.net/def/waterml/2.0/interpolationType/Continuous
+         */
+        Continuous,
+        /**
+         * http://www.opengis.net/def/waterml/2.0/interpolationType/Discontinuous
+         */
+        Discontinuous,
+        /**
+         * Instantaneous total
+         * http://www.opengis.net/def/waterml/2.0/interpolationType/InstantTotal
+         */
+        InstantTotal,
+        /**
+         * Average in preceding interval
+         * http://www.opengis.net/def/waterml/2.0/interpolationType/AveragePrec
+         */
+        AveragePrec,
+        /**
+         * Maximum in preceding interval
+         * http://www.opengis.net/def/waterml/2.0/interpolationType/MaxPrec
+         */
+        MaxPrec,
+        /**
+         * Minimum in preceding interval
+         * http://www.opengis.net/def/waterml/2.0/interpolationType/MinPrec
+         */
+        MinPrec,
+        /**
+         * Preceding total
+         * http://www.opengis.net/def/waterml/2.0/interpolationType/TotalPrec
+         */
+        TotalPrec,
+        /**
+         * Average in succeeding interval
+         * http://www.opengis.net/def/waterml/2.0/interpolationType/AverageSucc
+         */
+        AverageSucc,
+        /**
+         * Succeeding total
+         * http://www.opengis.net/def/waterml/2.0/interpolationType/TotalSucc
+         */
+        TotalSucc,
+        /**
+         * Minimum in succeeding interval
+         * http://www.opengis.net/def/waterml/2.0/interpolationType/MinSucc
+         */
+        MinSucc,
+        /**
+         * Maximum in succeeding interval
+         * http://www.opengis.net/def/waterml/2.0/interpolationType/MaxSucc
+         */
+        MaxSucc,
+        /**
+         * Constant in preceding interval
+         * http://www.opengis.net/def/waterml/2.0/interpolationType/ConstPrec
+         */
+        ConstPrec,
+        /**
+         * Constant in succeeding interval
+         * http://www.opengis.net/def/waterml/2.0/interpolationType/ConstSucc
+         */
+        ConstSucc,
+        /**
+         * Statistical
+         * http://www.opengis.net/def/waterml/2.0/interpolationType/Statistical
+         */
+        Statistical;
+
+        public String getIdentifier() {
+            return INTERPOLATION_TYPE + "/" + this.toString();
+        }
+
+        public String getTitle() {
+            return this.toString();
+        }
+
+    }
+
+    
 }
