+++ hibernate/common/src/main/java/org/n52/sos/ds/hibernate/util/observation/WaterMLMetadataAdder.java
@@ -33,12 +33,12 @@ import org.n52.sos.exception.CodedException;
 import org.n52.sos.exception.ows.NoApplicableCodeException;
 import org.n52.sos.ogc.om.OmObservation;
 import org.n52.sos.ogc.om.OmObservationConstellation;
-import org.n52.sos.ogc.wml.DefaultPointMetadata;
-import org.n52.sos.ogc.wml.DefaultTVPMeasurementMetadata;
-import org.n52.sos.ogc.wml.MeasurementTimeseriesMetadata;
-import org.n52.sos.ogc.wml.Metadata;
-import org.n52.sos.ogc.wml.WaterMLConstants;
-import org.n52.sos.ogc.wml.WaterMLConstants.InterpolationType;
+import org.n52.sos.ogc.series.wml.DefaultPointMetadata;
+import org.n52.sos.ogc.series.wml.DefaultTVPMeasurementMetadata;
+import org.n52.sos.ogc.series.wml.MeasurementTimeseriesMetadata;
+import org.n52.sos.ogc.series.wml.Metadata;
+import org.n52.sos.ogc.series.wml.WaterMLConstants;
+import org.n52.sos.ogc.series.wml.WaterMLConstants.InterpolationType;
 
 import com.google.common.base.Optional;
